# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Field Inspect API
  version: 0.1.0
  description: Field Inspect offline inspection API. Local loopback access by default;
    configured credentials required for network access. Processing completion is not
    field acceptance. No vehicle control or online package-upload endpoint. New functions
    have not been tested.
paths:
...
  /api/instruments/assets:
    get:
      tags:
      - Instruments
      summary: Assets
      operationId: assets_api_instruments_assets_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
      - Instruments
      summary: Register
      operationId: register_api_instruments_assets_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Asset'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/records/{kind}:
    get:
      tags:
      - Instruments
      summary: List Records
      operationId: list_records_api_instruments_records__kind__get
      parameters:
      - name: kind
        in: path
        required: true
        schema:
          enum:
          - observation
          - localization
          - asset_match
          - station_attempt
          - reading
          - match_review
          - station_evidence
          - processing_report
          type: string
          title: Kind
      - name: run_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Instruments
      summary: Put Record
      operationId: put_record_api_instruments_records__kind__post
      parameters:
      - name: kind
        in: path
        required: true
        schema:
          enum:
          - observation
          - localization
          - station_attempt
          type: string
          title: Kind
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/matches:
    post:
      tags:
      - Instruments
      summary: Match
      operationId: match_api_instruments_matches_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchInput'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/reviews:
    post:
      tags:
      - Instruments
      summary: Review
      operationId: review_api_instruments_reviews_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchReview'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/readings:
    post:
      tags:
      - Instruments
      summary: Reading
      operationId: reading_api_instruments_readings_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadingInput'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/history:
    post:
      tags:
      - Instruments
      summary: History
      operationId: history_api_instruments_history_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoryBatch'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/assets/{asset_id}/analysis:
    get:
      tags:
      - Instruments
      summary: Analysis
      operationId: analysis_api_instruments_assets__asset_id__analysis_get
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      - name: at
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: At
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/observations/{observation_id}/analysis:
    get:
      tags:
      - Instruments
      summary: Observed Analysis
      operationId: observed_analysis_api_instruments_observations__observation_id__analysis_get
      parameters:
      - name: observation_id
        in: path
        required: true
        schema:
          type: string
          title: Observation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/runs/{run_id}/reindex-stations:
    post:
      tags:
      - Instruments
      summary: Reindex Stations
      operationId: reindex_stations_api_instruments_runs__run_id__reindex_stations_post
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/stations/{attempt_id}/captures/{sequence}/process:
    post:
      tags:
      - Instruments
      summary: Process Capture
      operationId: process_capture_api_instruments_stations__attempt_id__captures__sequence__process_post
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: string
          title: Attempt Id
      - name: sequence
        in: path
        required: true
        schema:
          type: integer
          title: Sequence
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/pages/{kind}:
    get:
      tags:
      - Instruments
      summary: Record Page
      operationId: record_page_api_instruments_pages__kind__get
      parameters:
      - name: kind
        in: path
        required: true
        schema:
          enum:
          - observation
          - localization
          - asset_match
          - station_attempt
          - reading
          - match_review
          - station_evidence
          - processing_report
          type: string
          title: Kind
      - name: run_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Run Id
      - name: attempt_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Attempt Id
      - name: cursor
        in: query
        required: false
        schema:
          type: string
          maxLength: 512
          default: ''
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/records/{kind}/{identity}:
    get:
      tags:
      - Instruments
      summary: Get Record
      operationId: get_record_api_instruments_records__kind___identity__get
      parameters:
      - name: kind
        in: path
        required: true
        schema:
          enum:
          - observation
          - localization
          - asset_match
          - station_attempt
          - reading
          - match_review
          - station_evidence
          - processing_report
          type: string
          title: Kind
      - name: identity
        in: path
        required: true
        schema:
          type: string
          title: Identity
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/runs/{run_id}/jobs:
    post:
      tags:
      - Instruments
      summary: Create Job
      operationId: create_job_api_instruments_runs__run_id__jobs_post
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/jobs:
    get:
      tags:
      - Instruments
      summary: List Jobs
      operationId: list_jobs_api_instruments_jobs_get
      parameters:
      - name: run_id
        in: query
        required: true
        schema:
          type: string
          title: Run Id
      - name: cursor
        in: query
        required: false
        schema:
          type: string
          maxLength: 128
          default: ''
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/jobs/{job_id}:
    get:
      tags:
      - Instruments
      summary: Get Job
      operationId: get_job_api_instruments_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/jobs/{job_id}/{action}:
    post:
      tags:
      - Instruments
      summary: Control Job
      operationId: control_job_api_instruments_jobs__job_id___action__post
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: action
        in: path
        required: true
        schema:
          enum:
          - pause
          - resume
          - retry_failed
          type: string
          title: Action
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/jobs/{job_id}/items:
    get:
      tags:
      - Instruments
      summary: Job Items
      operationId: job_items_api_instruments_jobs__job_id__items_get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: cursor
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/jobs/{job_id}/events:
    get:
      tags:
      - Instruments
      summary: Job Events
      operationId: job_events_api_instruments_jobs__job_id__events_get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: cursor
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/asset-page:
    get:
      tags:
      - Instruments
      summary: Asset Page
      operationId: asset_page_api_instruments_asset_page_get
      parameters:
      - name: scene_version_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Scene Version Id
      - name: cursor
        in: query
        required: false
        schema:
          type: string
          maxLength: 128
          default: ''
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/observations/{observation_id}/review-context:
    get:
      tags:
      - Instruments
      summary: Review Context
      operationId: review_context_api_instruments_observations__observation_id__review_context_get
      parameters:
      - name: observation_id
        in: path
        required: true
        schema:
          type: string
          title: Observation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/observations/{observation_id}/match:
    post:
      tags:
      - Instruments
      summary: Match Observed
      operationId: match_observed_api_instruments_observations__observation_id__match_post
      parameters:
      - name: observation_id
        in: path
        required: true
        schema:
          type: string
          title: Observation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Asset:
      properties:
        asset_id:
          type: string
          maxLength: 128
          minLength: 1
          title: Asset Id
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        scene_version_id:
          type: string
          minLength: 1
          title: Scene Version Id
        position:
          $ref: '#/components/schemas/Position3D'
        marker_family:
          type: string
          minLength: 1
          title: Marker Family
        marker_id:
          type: integer
          minimum: 0.0
          title: Marker Id
        marker_role:
          type: string
          const: instrument_proxy
          title: Marker Role
        match_tolerance_m:
          type: number
          exclusiveMinimum: 0.0
          title: Match Tolerance M
        units:
          additionalProperties:
            type: string
          propertyNames:
            enum:
            - deep_base
            - shallow_base
            - delta
          type: object
          title: Units
        thresholds:
          additionalProperties:
            additionalProperties:
              type: number
            type: object
          propertyNames:
            enum:
            - deep_base
            - shallow_base
            - delta
          type: object
          title: Thresholds
          default: {}
        rule_version:
          type: string
          minLength: 1
          title: Rule Version
      additionalProperties: false
      type: object
      required:
      - asset_id
      - name
      - scene_version_id
      - position
      - marker_family
      - marker_id
      - marker_role
      - match_tolerance_m
      - units
      - rule_version
      title: Asset
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HistoryBatch:
      properties:
        run_id:
          type: string
          minLength: 1
          title: Run Id
        mapping:
          $ref: '#/components/schemas/HistoryMapping'
        source_type:
          type: string
          enum:
          - replay
          - simulation
          title: Source Type
        rows:
          items:
            additionalProperties: true
            type: object
          type: array
          maxItems: 10000
          minItems: 1
          title: Rows
      additionalProperties: false
      type: object
      required:
      - run_id
      - mapping
      - source_type
      - rows
      title: HistoryBatch
...
    MatchInput:
      properties:
        localization_id:
          type: string
          title: Localization Id
        marker_family:
          type: string
          title: Marker Family
        marker_id:
          type: integer
          minimum: 0.0
          title: Marker Id
      additionalProperties: false
      type: object
      required:
      - localization_id
      - marker_family
      - marker_id
      title: MatchInput
    MatchReview:
      properties:
        review_id:
          type: string
          minLength: 1
          title: Review Id
        match_id:
          type: string
          minLength: 1
          title: Match Id
        asset_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Asset Id
        operator_label:
          type: string
          maxLength: 100
          minLength: 1
          title: Operator Label
        reason:
          type: string
          maxLength: 1000
          minLength: 1
          title: Reason
        expected_review_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Expected Review Id
      additionalProperties: false
      type: object
      required:
      - review_id
      - match_id
      - asset_id
      - operator_label
      - reason
      title: MatchReview
...
    Position3D:
      properties:
        x:
          type: number
          title: X
        y:
          type: number
          title: Y
        z:
          type: number
          title: Z
      additionalProperties: false
      type: object
      required:
      - x
      - y
      - z
      title: Position3D
...
    ReadingInput:
      properties:
        reading:
          $ref: '#/components/schemas/InstrumentReading'
        mapping_version:
          type: string
          minLength: 1
          title: Mapping Version
        source_timezone:
          type: string
          minLength: 1
          title: Source Timezone
        source_reference:
          type: string
          minLength: 1
          title: Source Reference
        raw_fields:
          additionalProperties: true
          type: object
          title: Raw Fields
      additionalProperties: false
      type: object
      required:
      - reading
      - mapping_version
      - source_timezone
      - source_reference
      - raw_fields
      title: ReadingInput
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    OperatorSession:
      type: apiKey
      in: cookie
      name: astra_session
servers:
- url: /
security:
- BearerAuth: []
- OperatorSession: []